requirements.txt
discord.py>=1.7.3
unicodedata2>=14.0.0
aiohttp>=3.8.1